Osc: a falling timbre sweep fell nowhere
kAudioBlockSize is size_t, so the ramp's divisor was unsigned and dragged the
numerator unsigned with it. kFalling is the one sweep whose numerator is
negative, and division does not survive that conversion the way the truncation
to int16 does -- so the falling ramp came out as noise hovering near its
starting value instead of descending.

Measured on VARIABLE_PULSE at MIDI 48, pulse width over six segments:
  rising   104-170  61-101  37-60  21-36  12-20   4-11
  falling  168-172  172-172 101-172 172-172 29-172 165-172   before
  falling    5-11    11-20  21-35  36-61  63-101  97-172     after
The falling case is now the rising one's mirror.

ubsan.cc had it too, where `to` is below `from` on every other block: half its
ramps were noise, so the values that sweep exists to walk were never walked.

Every golden moved, because every shape renders the falling case. Re-recorded:
they pinned a corrupted input ramp.

+++ tools/osctest/driver.cc
@@ -204,8 +204,14 @@ uint32_t HashShape(int shape, bool dump) {
         SweepRange(sweep, &from, &to);
         const int16_t raw_timbre = g_hold_timbre
             ? static_cast<int16_t>(g_held_timbre)
+            // Signed divisor: kAudioBlockSize is size_t, so an unsigned
+            // divisor drags the numerator unsigned with it, and a FALLING
+            // sweep's numerator is negative. Division does not survive that
+            // the way the truncation to int16 survives it, so the falling
+            // ramp came out as noise around its starting value.
             : static_cast<int16_t>(
-                from + (to - from) * step / (g_blocks * kAudioBlockSize));
+                from + (to - from) * step /
+                    static_cast<long>(g_blocks * kAudioBlockSize));
         timbre_gain[i] = g_warp_timbre
             ? osc.WarpTimbre(raw_timbre, static_cast<OscillatorShape>(shape))
             : raw_timbre;
